+++ src/pluck.c
@@ -50,13 +50,15 @@ int find_offset(SEXP x, SEXP index, int i, bool strict) {
       goto numeric_index_error;
     }
 
-    --val;
+    if (val < 0) {
+      val = n + val + 1;
+    }
     if (check_double_index_length(val, n, i, strict)) {
       goto numeric_index_error;
     }
 
     UNPROTECT(n_protect);
-    return val;
+    return val - 1;
 
    numeric_index_error:
     UNPROTECT(n_protect);
@@ -93,7 +95,7 @@ int find_offset(SEXP x, SEXP index, int i, bool strict) {
 
     }
     if (strict) {
-      Rf_errorcall(R_NilValue, "Can't find name `%s` in vector", val);
+      Rf_errorcall(R_NilValue, "Can't find name `%s` in vector.", val);
     } else {
       UNPROTECT(1);
       return -1;
@@ -130,7 +132,7 @@ SEXP extract_vector(SEXP x, SEXP index_i, int i, bool strict) {
   case RAWSXP:  return Rf_ScalarRaw(RAW(x)[offset]) ;
   default:
     Rf_errorcall(R_NilValue,
-      "Don't know how to index object of type %s at level %d",
+      "Don't know how to index object of type %s at level %d.",
       Rf_type2char(TYPEOF(x)), i + 1
     );
   }
@@ -225,7 +227,7 @@ SEXP pluck_impl(SEXP x, SEXP index, SEXP missing, SEXP strict_arg) {
     switch (TYPEOF(x)) {
     case NILSXP:
       if (strict) {
-        Rf_errorcall(R_NilValue, "Plucked object can't be NULL");
+        Rf_errorcall(R_NilValue, "Plucked object can't be NULL.");
       }
       // Leave the indexing loop early
       goto end;
@@ -249,7 +251,7 @@ SEXP pluck_impl(SEXP x, SEXP index, SEXP missing, SEXP strict_arg) {
       REPROTECT(x, idx);
       break;
     default:
-      Rf_errorcall(R_NilValue, "Can't pluck from a %s", Rf_type2char(TYPEOF(x)));
+      Rf_errorcall(R_NilValue, "Can't pluck from a %s.", Rf_type2char(TYPEOF(x)));
     }
 
   }
@@ -267,7 +269,7 @@ static int check_input_lengths(int n, SEXP index, int i, bool strict) {
 
   if (n == 0) {
     if (strict) {
-      Rf_errorcall(R_NilValue, "Plucked object must have at least one element");
+      Rf_errorcall(R_NilValue, "Plucked object must have at least one element.");
     } else {
       return -1;
     }
@@ -287,7 +289,7 @@ static int check_double_index_finiteness(double val, SEXP index, int i, bool str
 
   if (strict) {
     Rf_errorcall(R_NilValue,
-                 "Index %d must be finite, not %s",
+                 "Index %d must be finite, not %s.",
                  i + 1,
                  Rf_translateCharUTF8(Rf_asChar(index)));
   } else {
@@ -296,21 +298,29 @@ static int check_double_index_finiteness(double val, SEXP index, int i, bool str
 }
 
 static int check_double_index_length(double val, int n, int i, bool strict) {
-  if (val < 0) {
+  if (val == 0) {
     if (strict) {
-      Rf_errorcall(R_NilValue,
-                   "Index %d must be greater than 0, not %.0f",
-                   i + 1,
-                   val + 1);
+      Rf_errorcall(R_NilValue, "Index %d is zero.", i + 1);
     } else {
       return -1;
     }
-  } else if (val >= n) {
+  } else if (val < 0) {
+    if (strict) {
+      // Negative values have already been subtracted from end
+      Rf_errorcall(R_NilValue,
+                   "Negative index %d must be greater than or equal to %d, not %.0f.",
+                   i + 1,
+                   -n,
+                   val - n - 1);
+    } else {
+      return -1;
+    }
+  } else if (val > n) {
     if (strict) {
       Rf_errorcall(R_NilValue,
-                   "Index %d exceeds the length of plucked object (%.0f > %d)",
+                   "Index %d exceeds the length of plucked object (%.0f > %d).",
                    i + 1,
-                   val + 1,
+                   val,
                    n);
     } else {
       return -1;
@@ -323,7 +333,7 @@ static int check_double_index_length(double val, int n, int i, bool strict) {
 static int check_character_index(SEXP string, int i, bool strict) {
   if (string == NA_STRING) {
     if (strict) {
-      Rf_errorcall(R_NilValue, "Index %d can't be NA", i + 1);
+      Rf_errorcall(R_NilValue, "Index %d can't be NA.", i + 1);
     } else {
       return -1;
     }
@@ -333,7 +343,7 @@ static int check_character_index(SEXP string, int i, bool strict) {
   const char* val = CHAR(string);
   if (val[0] == '\0') {
     if (strict) {
-      Rf_errorcall(R_NilValue, "Index %d can't be an empty string (\"\")", i + 1);
+      Rf_errorcall(R_NilValue, "Index %d can't be an empty string (\"\").", i + 1);
     } else {
       return -1;
     }
@@ -348,7 +358,7 @@ static int check_names(SEXP names, int i, bool strict) {
   }
 
   if (strict) {
-    Rf_errorcall(R_NilValue, "Index %d is attempting to pluck from an unnamed vector using a string name", i + 1);
+    Rf_errorcall(R_NilValue, "Index %d is attempting to pluck from an unnamed vector using a string name.", i + 1);
   } else {
     return -1;
   }
@@ -361,7 +371,7 @@ static int check_offset(int offset, SEXP index_i, bool strict) {
 
   if (strict) {
     Rf_errorcall(R_NilValue,
-                 "Can't find index `%s` in vector",
+                 "Can't find index `%s` in vector.",
                  Rf_translateCharUTF8(Rf_asChar(index_i)));
   } else {
     return -1;
@@ -375,7 +385,7 @@ static int check_unbound_value(SEXP val, SEXP index_i, bool strict) {
 
   if (strict) {
     Rf_errorcall(R_NilValue,
-                 "Can't find object `%s` in environment",
+                 "Can't find object `%s` in environment.",
                  Rf_translateCharUTF8(Rf_asChar(index_i)));
   } else {
     return -1;
@@ -399,7 +409,7 @@ static int check_s4_slot(SEXP val, SEXP index_i, bool strict) {
 static int check_obj_length(SEXP n, bool strict) {
   if (TYPEOF(n) != INTSXP || Rf_length(n) != 1) {
     if (strict) {
-      Rf_errorcall(R_NilValue, "Length of S3 object must be a scalar integer");
+      Rf_errorcall(R_NilValue, "Length of S3 object must be a scalar integer.");
     } else {
       return -1;
     }